Add WASM support
Moved some functions and constants around so common ones are in
`ui_common.js`.

`ui_js.go` Mostly contains a basic call out to `prompt()` to ask for
input. It's not perfect, but it works and is simple.
