+++ ui_colored.go
@@ -1,5 +1,7 @@
 // Copyright (c) HashiCorp, Inc.
 // SPDX-License-Identifier: MPL-2.0
+//go:build !js
+// +build !js
 
 package cli
