+++ ui_js.go
@@ -1,0 +1,10 @@
+package cli
+
+import (
+	"syscall/js"
+)
+
+func (u *BasicUi) ask(query string, secret bool) (string, error) {
+	line := js.Global().Call("prompt", query).String()
+	return line, nil
+}
